# reduce memory use by lexer