```console
$ benchstat a.txt b.txt                             
goos: darwin
goarch: arm64
pkg: go.expect.digital/mf2/parse
cpu: Apple M2 Max
                         │    a.txt    │                b.txt                │
                         │   sec/op    │   sec/op     vs base                │
ComplexMessage_String-12   4.169µ ± 1%   4.176µ ± 2%        ~ (p=0.172 n=10)
Lex-12                     2.006µ ± 0%   1.405µ ± 0%  -29.96% (p=0.000 n=10)
Parse-12                   3.833µ ± 0%   3.342µ ± 1%  -12.82% (p=0.000 n=10)
geomean                    3.177µ        2.696µ       -15.12%

                         │    a.txt     │                 b.txt                  │
                         │     B/op     │     B/op      vs base                  │
ComplexMessage_String-12   27.16Ki ± 0%   27.16Ki ± 0%        ~ (p=1.000 n=10) ¹
Lex-12                       856.0 ± 0%     656.0 ± 0%  -23.36% (p=0.000 n=10)
Parse-12                   7.438Ki ± 0%   7.242Ki ± 0%   -2.63% (p=0.000 n=10)
geomean                    5.528Ki        5.014Ki        -9.30%
¹ all samples are equal

                         │   a.txt    │                b.txt                 │
                         │ allocs/op  │ allocs/op   vs base                  │
ComplexMessage_String-12   61.00 ± 0%   61.00 ± 0%        ~ (p=1.000 n=10) ¹
Lex-12                     43.00 ± 0%   19.00 ± 0%  -55.81% (p=0.000 n=10)
Parse-12                   77.00 ± 0%   53.00 ± 0%  -31.17% (p=0.000 n=10)
geomean                    58.67        39.46       -32.75%
¹ all samples are equal
```